FIX: Transparancy problems in ZERNSimpleMaterial has been fixed. FIX: ZEModelMeshLOD DrawEndDistance property integer overflow has been fixed. FIX: ZERNMaterial::Update Initialization problem has been fixed.
svn path=/branches/v0.6.5-Editor/; revision=2025

+++ Source/ZEEngine/ZEModel/ZEModelMesh.cpp
@@ -751,7 +751,7 @@ bool ZEModelMesh::PreRender(const ZERNPreRenderParameters* Parameters)
 		}
 	}
 
-	if (ClosestBoundingBoxEdgeDistanceSquare > (LODs[CurrentLOD].GetDrawEndDistance() * LODs[CurrentLOD].GetDrawEndDistance()))
+	if (ClosestBoundingBoxEdgeDistanceSquare > ((float)LODs[CurrentLOD].GetDrawEndDistance() * (float)LODs[CurrentLOD].GetDrawEndDistance()))
 		return false;
 
 	ZEModelMeshLOD* MeshLOD = &LODs[(ZESize)CurrentLOD];
